Update for lumberjack 2
