+++ CHANGE_LOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.0.0
+
+### Changed
+
+- Requires lumberjack >= 2.0
+- **Breaking Change** Renamed `tags` to `attributes` in the ECS mapping.
+
+### Removed
+
+- Removed support for Ruby < 2.7.
+
 ## 1.0.0
 
 ### Added
