+++ Gemfile
@@ -2,6 +2,9 @@ source "https://rubygems.org"
 
 gemspec
 
+gem "lumberjack", github: "bdurand/lumberjack", branch: "lumberjack_2"
+gem "lumberjack_json_device", github: "bdurand/lumberjack_json_device", branch: "lumberjack-2-dependency"
+
 gem "dotenv"
 gem "rspec", "~> 3.10"
 gem "appraisal"
