+++ spec/lumberjack/ecs_device_spec.rb
